requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
numpy==1.24.3
tensorflow==2.13.0
librosa==0.10.1
scikit-learn==1.3.0
Pillow==10.0.0
sounddevice==0.4.6
matplotlib==3.7.2
scipy==1.11.2
numba==0.57.1
